+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farkle/core
 
+## 1.2.1
+
+### Patch Changes
+
+- Add vitest tests
+
 ## 1.2.0
 
 ### Minor Changes